Rename components to services (#350)
