+++ code_snippets/java/src/main/java/concepts/services/RoleUpdateService.java
@@ -1,4 +1,4 @@
-package concepts.components;
+package concepts.services;
 
 import dev.restate.sdk.Context;
 import dev.restate.sdk.annotation.Handler;
