+++ code_snippets/java/src/main/java/concepts/services/SystemB.java
@@ -1,4 +1,4 @@
-package concepts.components;
+package concepts.services;
 
 public class SystemB {
     public static boolean applyPermission(String userId, String permission) {
