+++ code_snippets/java/src/main/java/concepts/services/UpdateRequest.java
@@ -1,4 +1,4 @@
-package concepts.components;
+package concepts.services;
 
 public class UpdateRequest {
     private String userId;
