+++ docs/concepts/invocations.mdx
@@ -114,11 +114,11 @@ Invocations get a unique identifier. This identifier is used to track the progre
             <CH.Code>
 
                 ```java RestateService.java
-                CODE_LOAD::java/src/main/java/concepts/components/invocations/RpcCalls.java#<start_rpc>-<end_rpc>
+                CODE_LOAD::java/src/main/java/concepts/invocations/RpcCalls.java#<start_rpc>-<end_rpc>
                 ```
 
                 ```java PlainJavaService.java
-                CODE_LOAD::java/src/main/java/concepts/components/invocations/RpcCalls.java#<start_rpc_java>-<end_rpc_java>
+                CODE_LOAD::java/src/main/java/concepts/invocations/RpcCalls.java#<start_rpc_java>-<end_rpc_java>
                 ```
 
             </CH.Code>
@@ -133,11 +133,11 @@ Invocations get a unique identifier. This identifier is used to track the progre
             <CH.Code>
 
                 ```java RestateService.java
-                CODE_LOAD::java/src/main/java/concepts/components/invocations/OneWayCalls.java#<start_one_way_call>-<end_one_way_call>
+                CODE_LOAD::java/src/main/java/concepts/invocations/OneWayCalls.java#<start_one_way_call>-<end_one_way_call>
                 ```
 
                 ```java PlainJavaService.java
-                CODE_LOAD::java/src/main/java/concepts/components/invocations/OneWayCalls.java#<start_one_way_call_java>-<end_one_way_call_java>
+                CODE_LOAD::java/src/main/java/concepts/invocations/OneWayCalls.java#<start_one_way_call_java>-<end_one_way_call_java>
                 ```
 
 
@@ -151,11 +151,11 @@ Invocations get a unique identifier. This identifier is used to track the progre
             <CH.Code>
 
                 ```java RestateService.java
-                CODE_LOAD::java/src/main/java/concepts/components/invocations/DelayedCalls.java#<start_delayed_call>-<end_delayed_call>
+                CODE_LOAD::java/src/main/java/concepts/invocations/DelayedCalls.java#<start_delayed_call>-<end_delayed_call>
                 ```
 
                 ```java PlainJavaService.java
-                CODE_LOAD::java/src/main/java/concepts/components/invocations/DelayedCalls.java#<start_delayed_call_java>-<end_delayed_call_java>
+                CODE_LOAD::java/src/main/java/concepts/invocations/DelayedCalls.java#<start_delayed_call_java>-<end_delayed_call_java>
                 ```
 
             </CH.Code>
