+++ docs/develop/java/serving.mdx
@@ -77,7 +77,7 @@ The implementation of your service methods remains the same for both deployment 
 
 ## Using Java 21 Virtual Threads
 
-If you use a JVM >= 21, you can easily use virtual threads to run your components:
+If you use a JVM >= 21, you can easily use virtual threads to run your services:
 
 <Tabs groupId="sdk" queryString>
 <TabItem value="java" label="Java">
